feature: :✨:  ディレクトリ構造をsrc/appの形式に修正しました
